+++ types/components/dv-autocomplete.locale.d.ts
@@ -1,0 +1,10 @@
+declare namespace _default {
+    namespace en {
+        let label: string;
+    }
+    namespace tr {
+        let label_1: string;
+        export { label_1 as label };
+    }
+}
+export default _default;