+++ types/components/dv-product-card.locale.d.ts
@@ -1,0 +1,10 @@
+declare namespace _default {
+    namespace en {
+        let action: string;
+    }
+    namespace tr {
+        let action_1: string;
+        export { action_1 as action };
+    }
+}
+export default _default;