+++ types/components/dv-toast-stack.locale.d.ts
@@ -1,0 +1,13 @@
+declare namespace _default {
+    namespace en {
+        let label: string;
+        let dismiss: string;
+    }
+    namespace tr {
+        let label_1: string;
+        export { label_1 as label };
+        let dismiss_1: string;
+        export { dismiss_1 as dismiss };
+    }
+}
+export default _default;